refac: rename the existing tests adding adding the Azure IoT Hub suffix
Changelog: None
Ticket: MEN-5451
