+++ tests/mmock/config/test_sync_azure_iothub/deviceauth_get_devices_TenantSync01.yml
@@ -2,7 +2,7 @@ request:
   description: Call to device auth for tenant "TenantSync01"
   host: mender-device-auth
   method: GET
-  path: /api/internal/v1/devauth/tenants/TestSync01/devices
+  path: /api/internal/v1/devauth/tenants/TestSyncAzureIoTHub01/devices
   queryStringParameters:
     per_page: ["10"]
     id:
