+++ tests/mmock/config/test_sync_azure_iothub/iothub_delete_device_7abb6133-ad97-44ba-a159-674242ee565e.yml
