+++ tests/mmock/config/test_sync_azure_iothub/iothub_delete_device_9b2083e4-83b6-41d3-b089-4d2ec137620b.yml
