+++ tests/mmock/config/test_sync_azure_iothub/iothub_get_device_966095ec-6bdd-4a76-8498-3c0dffdb9ee2.yml
