+++ tests/mmock/config/test_sync_azure_iothub/iothub_patch_twins_93406e21-8e3f-4435-9786-a294a70298ee.yml
