+++ tests/mmock/config/test_sync_azure_iothub/iothub_put_device_1424a387-3431-425b-9f44-1c8eba21812d.yml
