+++ tests/mmock/config/test_sync_azure_iothub/iothub_put_device_93406e21-8e3f-4435-9786-a294a70298ee.yml
