+++ tests/mmock/config/test_sync_azure_iothub/workflows_provision_external_device_93406e21-8e3f-4435-9786-a294a70298ee.yml
